sqlite: upgrade to SQLite 3.13.0
Downloaded from http://sqlite.com/2016/sqlite-amalgamation-3130000.zip

  $ sha256sum sqlite-amalgamation-3130000.zip
  f9ef8811f5f221d81a3ba811a8dfdcc54b19acc6d8032c74719ff95ed0026611  sqlite-amalgamation-3130000.zip

dist/orig contains the stock sqlite3 code, as packaged in the ZIP file above.

dist contains a copy of dist/orig, but with the Android.patch file applied.
Please see Android.patch for a list of differences between stock and
Android.

The Android specific patch for sqlite3.c failed to apply cleanly.
The conflicts were resolved and the Android.patch file was regenerated.

It is to note that the Android specific changes from "off_t" to "off64_t" for "osPread64" and "osPwrite64" are no longer necessary.
Upstream applied these changes in the newest 3.13.0 release.

Also this patch deletes the dist/version file as it is obsoleted and replaced by README.version for Android.

The changes in the new releases can be found here: https://www.sqlite.org/changes.html

Change-Id: I20001055339015b3c05836f7cef7b724700b08a9
Signed-off-by: Alex Naidis <alex.naidis@linux.com>